requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# AI/ML Libraries
google-genai>=1.0.0
aiohttp>=3.8.0

# Stock Data & Analysis
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0

# Data Processing & Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# File Generation
jinja2>=3.1.0
markdown>=3.5.0

# Logging & Error Handling
rich>=13.0.0